Split migration related functions to new module `migration.py`
Create new module `load_contracts.py` to load contracts as it is
required in both migration and core.
